==> 08-eof.Rmd <==

# 끝 맺으며 {#eof}


지금까지 중학교 $1$학년 수학 교과 과정에 나오는 내용들을 **R** 사용자 관점에서 재구성해 보았습니다. 그때 그때 필요한 개념들을 다루었기 때문에 **R**의 개념들을 차근 차근 순서대로 다루는 데에는 한계가 있었습니다. 다만, 중학교 수학 교과 과정에서 다루는 개념과 내용들을 R과 같은 데이터 과학 도구를 이용하여 구현해 보았다는 데 의미를 두고 싶습니다.

이 책에 나온 내용을 여러분들이 전부 다 이해할 필요는 없습니다. 다만, 교과 과정 뿐만 아니라 교과 이외의 내용들에도 관심을 가지고 하나씩 알아나가고 또 조금씩 깊이를 더해 나가면 멀지 않은 미래에 여러분이 흥미를 느끼는 분야의 깊이 있는 지식을 가지게 될 것입니다. 이 책이 여러분들이 기본 수학 개념을 이해하는데 조금 도움이 되고 데이터 과학에 흥미를 느끼는 계기가 되면 좋겠습니다.
